+++ lbz/handlers.py
@@ -2,6 +2,7 @@ from abc import ABCMeta, abstractmethod
 from typing import Generic, Optional, TypeVar
 
 from lbz.misc import deprecated, get_logger
+from lbz.type_defs import LambdaContext
 
 logger = get_logger(__name__)
 
@@ -9,7 +10,7 @@ T = TypeVar("T")
 
 
 class BaseHandler(Generic[T], metaclass=ABCMeta):
-    def __init__(self, event: dict, context: object) -> None:
+    def __init__(self, event: dict, context: LambdaContext) -> None:
         self.raw_event = event
         self.context = context
         self.response: Optional[T] = None