+++ tests/test_lambdas/test_broker.py
@@ -4,6 +4,7 @@ from pytest import LogCaptureFixture
 
 from lbz.exceptions import NotFound
 from lbz.lambdas import LambdaBroker, LambdaResponse, LambdaResult, lambda_ok_response
+from lbz.type_defs import LambdaContext
 
 
 def simple_func(_data: dict = None) -> LambdaResponse:
@@ -18,24 +19,11 @@ class TestEventBroker:
         mapper = {"x": func}
         event = {"op": "x", "data": {"y": 1}}
 
-        resp = LambdaBroker(mapper, event).react()
+        resp = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert resp == {
             "result": LambdaResult.OK,
             "data": {"some": "data"},
-        }
-
-    def test_broker_works_properly_when_no_data_is_provided(self) -> None:
-        def no_data_func() -> LambdaResponse:
-            return lambda_ok_response()
-
-        mapper = {"x": no_data_func}
-        event = {"op": "x", "data": None}
-
-        resp = LambdaBroker(mapper, event).react()
-
-        assert resp == {
-            "result": LambdaResult.OK,
         }
 
     def test_broker_responds_with_contract_error_when_op_is_not_recognized(
@@ -44,7 +32,7 @@ class TestEventBroker:
         mapper = {"x": simple_func}
         event = {"op": "y", "data": {"z": 1}}
 
-        response = LambdaBroker(mapper, event).react()
+        response = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert response == {
             "result": LambdaResult.CONTRACT_ERROR,
@@ -64,7 +52,7 @@ class TestEventBroker:
         mapper = {"x": simple_func}
         event = {"data": {"y": 1}}
 
-        resp = LambdaBroker(mapper, event).react()
+        resp = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert resp == {
             "result": LambdaResult.CONTRACT_ERROR,
@@ -86,7 +74,7 @@ class TestEventBroker:
 
         mapper = {"x": func}
         event = {"op": "x", "data": {"y": 1}}
-        resp = LambdaBroker(mapper, event).react()
+        resp = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert resp == {"result": LambdaResult.SERVER_ERROR, "message": NotFound.message}
         assert caplog.record_tuples == [
@@ -105,7 +93,7 @@ class TestEventBroker:
 
         mapper = {"x": func}
         event = {"op": "x", "data": {"y": 1}}
-        resp = LambdaBroker(mapper, event).react()
+        resp = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert resp == {
             "result": LambdaResult.SERVER_ERROR,
@@ -126,7 +114,7 @@ class TestEventBroker:
 
         mapper = {"x": func}
         event = {"op": "x", "data": {"y": 1}}
-        resp = LambdaBroker(mapper, event).react()
+        resp = LambdaBroker(mapper, event, LambdaContext()).react()
 
         assert resp == {
             "result": LambdaResult.SERVER_ERROR,